Feedback and Reset functions

+++ lib/app/modules/bus/views/bus_view.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:nepal_express/app/models/seat.dart';
 import 'package:nepal_express/app/utils/constants.dart';
+import 'package:nepal_express/app/utils/memory.dart';
 import '../controllers/bus_controller.dart';
 
 class BusView extends StatelessWidget {
@@ -26,16 +28,20 @@ class BusView extends StatelessWidget {
               itemCount: controller.busesResponse?.buses?.length,
               itemBuilder: (context, index) {
                 var bus = controller.busesResponse!.buses![index];
-                return Card(
-                  elevation: 15,
-                  margin: const EdgeInsets.symmetric(
-                    horizontal: 12,
-                    vertical: 15,
-                  ),
-                  child: Container(
-                    decoration: BoxDecoration(
-                      color: Color.fromARGB(184, 215, 227, 233),
-                      borderRadius: BorderRadius.circular(15.0),
+                return GestureDetector(
+                  onTap: () {
+                    if (bus.id != null) {
+                      Get.to(() => BusDetailsView(), arguments: bus.id!);
+                    } else {
+                      // Handle case where busId is null
+                      print('Error: busId is null');
+                    }
+                  },
+                  child: Card(
+                    elevation: 5,
+                    margin: const EdgeInsets.symmetric(
+                      horizontal: 12,
+                      vertical: 15,
                     ),
                     child: Row(
                       children: [
@@ -59,13 +65,15 @@ class BusView extends StatelessWidget {
                                       fontWeight: FontWeight.bold,
                                     ),
                                   ),
-                                  Text(
-                                    bus.isDeleted == '1' ? ' (Deleted)' : '',
-                                    style: const TextStyle(
-                                      fontSize: 20,
-                                      color: Colors.red,
+                                  if (bus.isDeleted ==
+                                      '1') // Add this condition to show deleted text
+                                    Text(
+                                      ' (Deleted)',
+                                      style: const TextStyle(
+                                        fontSize: 20,
+                                        color: Colors.red,
+                                      ),
                                     ),
-                                  ),
                                 ],
                               ),
                               Text(
@@ -86,6 +94,8 @@ class BusView extends StatelessWidget {
                         const Spacer(),
                         IconButton(
                           onPressed: () {
+                            print(
+                                'isDeleted inside IconButton: ${bus.isDeleted}');
                             showDialog(
                               context: context,
                               builder: (context) {
@@ -99,7 +109,12 @@ class BusView extends StatelessWidget {
                                   actions: [
                                     TextButton(
                                       onPressed: () {
-                                        busController.deleteBus(bus.id ?? '');
+                                        Get.back();
+                                        if (bus.isDeleted == '0') {
+                                          busController.deleteBus(bus.id ?? '');
+                                        } else {
+                                          // Code to restore bus
+                                        }
                                       },
                                       child: const Text('Yes'),
                                     ),
@@ -115,12 +130,8 @@ class BusView extends StatelessWidget {
                             );
                           },
                           icon: bus.isDeleted == '1'
-                              ? const Icon(
-                                  Icons.restore,
-                                  color: Colors.green,
-                                )
-                              : const Icon(Icons.delete),
-                          color: Colors.red,
+                              ? const Icon(Icons.restore, color: Colors.green)
+                              : const Icon(Icons.delete, color: Colors.red),
                         ),
                       ],
                     ),
@@ -140,6 +151,73 @@ class BusView extends StatelessWidget {
               });
         },
         child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
+
+class BusDetailsView extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    final String busId = Get.arguments;
+    final BusController controller = Get.find();
+
+    // Call the getSeatsForBus method to fetch the seat data for the current busId
+    controller.getSeatsForBus(busId, Memory.getToken() ?? '');
+
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Bus Details'),
+      ),
+      body: SingleChildScrollView(
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              ElevatedButton(
+                onPressed: () {
+                  // Call the resetSeats function from the BusController
+                  controller.resetBusBookingStatus(busId);
+                },
+                child: Text('Reset Bus Booking Status'),
+              ),
+              SizedBox(height: 20),
+              ElevatedButton(
+                onPressed: () {
+                  // Call the resetSeats function from the BusController
+                  controller.resetSeats(busId);
+                },
+                child: Text('Reset Seats'),
+              ),
+              SizedBox(height: 20),
+              Text(
+                'Seat Data:',
+                style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+              ),
+              SizedBox(height: 20),
+              Obx(() {
+                if (controller.seats.isEmpty) {
+                  return Text('No seat data available.');
+                } else {
+                  return DataTable(
+                    columns: [
+                      DataColumn(label: Text('Seat Number')),
+                      DataColumn(label: Text('Availability')),
+                    ],
+                    rows: controller.seats.map((seat) {
+                      return DataRow(cells: [
+                        DataCell(Text(seat.seatNumber ?? '')),
+                        DataCell(Text(
+                            seat.availability == 1 ? 'Available' : 'Booked')),
+                      ]);
+                    }).toList(),
+                  );
+                }
+              }),
+            ],
+          ),
+        ),
       ),
     );
   }
@@ -333,10 +411,8 @@ class AddBusPopup extends StatelessWidget {
                       controller.refreshPage();
                     },
                     style: ElevatedButton.styleFrom(
-                      primary:
-                          Color.fromARGB(208, 62, 73, 111),
-                      onPrimary: const Color.fromARGB(
-                          255, 255, 255, 255),
+                      primary: Color.fromARGB(208, 62, 73, 111),
+                      onPrimary: const Color.fromARGB(255, 255, 255, 255),
                       elevation: 5,
                     ),
                     child: const Text('Add Bus'),
